+++ scijava/scijava-ops-indexer/src/main/java/module-info.java
@@ -1,4 +1,4 @@
-import org.scijava.ops.indexer.JavadocAnnotationProcessor;
+import org.scijava.ops.indexer.OpImplNoteParser;
 
 module org.scijava.javadoc.parser {
 	exports org.scijava.ops.indexer;
@@ -7,5 +7,5 @@ module org.scijava.javadoc.parser {
 	requires org.yaml.snakeyaml;
 	requires org.scijava.ops.spi;
 
-	provides javax.annotation.processing.Processor with JavadocAnnotationProcessor;
+	provides javax.annotation.processing.Processor with OpImplNoteParser;
 }
