+++ scijava/scijava-ops-indexer/src/main/resources/META-INF/services/javax.annotation.processing.Processor
@@ -1,1 +1,1 @@
-org.scijava.ops.indexer.JavadocAnnotationProcessor
+org.scijava.ops.indexer.OpImplNoteParser
